pyproject.toml
[tool.poetry]
name = "geochem-portal"
version = "0.1.0"
description = ""
authors = ["Edmond Chuc <edmond@kurrawong.ai>"]
readme = "README.md"
packages = [{include = "geochem_portal"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
rdflib = "^7.0.0"
uvicorn = "^0.23.2"
mangum = "^0.17.0"
httpx = "^0.25.0"
jsonschema = "^4.19.1"
pydantic-settings = "^2.0.3"
pyshacl = "^0.25.0"
python-multipart = "^0.0.9"
geochemxl = {url = "https://github.com/Kurrawong/geochemxl/archive/refs/tags/3.0.3.zip"}


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
pytest = "^7.4.1"
rich = "^13.5.2"
ruff = "^0.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = [
    "I001",
    "I002"
]